api_excercise/pyproject.toml
[tool.poetry]
name = "api-excercise"
version = "0.1.0"
description = ""
authors = ["Erasmo <erasmodimarco@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-cloud-bigquery = "^3.21.0"
google-cloud-storage = "^2.16.0"
mock = "^5.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
